requirements.txt
Django==2.2
django-model-utils==3.2.0
https://github.com/CeciliaSol/django-trix/archive/master.zip
git+https://github.com/fdemmer/django-weasyprint

